\name{allnans}
\alias{allnans}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Count all NAs/NaNs }
\description{
  Count NAs and NaNs in the data
}
\usage{
allnans(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a mtraix or a vector }
}

\value{
  \item{n}{number of NAs and NaNs}
}

\author{ Zornitsa Manolova }

\seealso{ See also \code{\link{allnotnans}}}
\keyword{ NA }
\keyword{ NaN }
\keyword{ count }
\keyword{ find }