\name{allnotnans}
\alias{allnotnans}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Count all numeric elements }
\description{
  Count all numeric elements in the data.
}
\usage{
allnotnans(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a matrix or vector}
}

\value{
  \item{n}{ number of all numeric elements}

}
\references{ ~put references to the literature/web site here ~ }
\author{ Zornitsa Manolova }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ NA }
\keyword{ NaN }
\keyword{ count }
\keyword{ find }