\name{countNaNs}
\alias{countNaNs}
\title{ Number of values equal to NaN along dimension d If d is omitted, it defaults to 1 (column-wise). }
\arguments{
\item{x}{data matrix}
\item{d}{dimension: (default) 1=column-wise 2=row-wise}
}
\value{
\item{n}{vector with number of values NaN}
}
\description{
 Number of values equal to NaN along dimension d If d is omitted, it defaults to 1 (column-wise). }
\author{Prof. Alfred Ultsch}
