\name{dbt.nanHandling-package}
\alias{dbt.nanHandling-package}
\alias{dbt.nanHandling}
\docType{package}
\title{
dbt.nanHandling-package
}
\description{
Handling of Not a Numbers (NaN)
}
\details{
\tabular{ll}{
Package: \tab dbt.nanHandling\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2016-08-22\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Zornitsa Manolova, Rabea Griese

Maintainer: Michael Thrun <mthrun@mathematik.uni-marburg.de>

}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

