\name{medianNaN}
\alias{medianNaN}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Median }
\description{
  Compute median value disregarding NaN values
}
\usage{
medianNaN(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a vector or a matrix }
}

\value{
  \item{m }{median value(s)}
}
\author{ Zornitsa Manolova }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ median }
