\name{nancorwithmean}
\alias{nancorwithmean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Substitute NA or NaN }
\description{
  The NaNs and NAs are substitute with meanvelues and then
  compute correlation
}
\usage{
nancorwithmean(Data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ a matrix }
}
\value{
  \item{cormat }{correlation matrix}
}
\author{ Zornitsa Manolova }

\seealso{\code{\link{nantomean}}}
\keyword{ correlation }
\keyword{ mean }
