\name{nanhmean}
\alias{nanhmean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Harmonic Mean }
\description{
  Compute harmonic mean value with NAs/NaNs and zero handling
}
\usage{
nanhmean(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a vector }
}

\value{
  \item{ m }{harmonic mean value}
}
\author{ Zornitsa Manolova }

\seealso{ \code{hmean} from the Package dbt.Statistics }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ mean }
