\name{nantozero}
\alias{nantozero}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ NaN to zero }
\description{
  Substitute NaNs and NAs in z by zero.
}
\usage{
nantozero(z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{ a vector or matrix }
}
\details{
  Use nantovalue.
}
\value{
  \item{ out }{ Data with NaNs replaced }
}

\author{ Zornitsa Manolova }
\seealso{  See Also as \code{\link{nantovalue}} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ NA }
\keyword{ NaN }
\keyword{ replace }
\keyword{ substitute }
