\name{removeNaN}
\alias{removeNaN}
\title{ remove rows with NaN elements from data  and remove corresponding elements from cls vector }
\arguments{
\item{Data}{data matrix or vector variable possibly containing NaN}
\item{Cls}{vector where the same elements are to be removed}
}
\value{cleanData   data matrix or vector cleaned of NaNs}
\description{
 remove rows with NaN elements from data  and remove corresponding elements from cls vector }
\author{
RG 04/15, imported from matlab}
